First Steps Towards Real Program.
This basically does the same thing as last time.
It just does it a little differently.

The actual "doing" has been moved out of the YACC file to schemaUpdate.c.
I wouldn't be surprised if the YACC file never needs to be touched again.
I wouldn't be surprised if it does, though, either.