Does What It's Supposed To.
Now it takes in a filename describing a DB schema on the command line, and a SQL dump on stdin and spits out the required lines to fit the data from the dump into this new DB.
It's a terrible bit of code.

It's poorly documented and it is probably a ton fragile.

I already can't really grasp how it works.

It also leaks like a sieve.

Oh well.
I'm sure I'll fix that at some point *wink*.